Rework `member.home` book search and implement borrowing restrictions

+++ app/Http/Controllers/BookController.php
@@ -44,6 +44,10 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
+        if(!\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+
         $validator = Validator::make($request->all(), [
             'title' => 'required|unique:books,title',
             'authors' => 'required|string|max:100',
@@ -85,6 +89,10 @@ class BookController extends Controller
      */
     public function show(Book $book)
     {
+        if(\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+
         $used_books = \App\Models\Transaction::where('book_id', '=', $book->id)
             ->where('status', '!=', 'returned')->count();
 
@@ -94,6 +102,18 @@ class BookController extends Controller
                 ->where('status', '!=', 'returned')
                 ->orderBy('date_to', 'ASC')->first();
         }
+
+        $book->hasPenalties = \App\Models\Penalty::where('transactions.user_id', '=', \Illuminate\Support\Facades\Auth::user()->id)
+            ->rightJoin('transactions', 'transactions.id', '=', 'penalties.transaction_id')
+            ->where('penalties.status', '=', 'unpaid')->count() 
+                ? 1 
+                : 0;
+
+        // select * from transactions where CURRENT_DATE() > date_to and user_id = 3 and status != 'returned'
+        $book->hasLateReturns = \App\Models\Transaction::whereRaw('CURRENT_DATE() > date_to')
+            ->where('user_id', \Illuminate\Support\Facades\Auth::user()->id)
+            ->where('status', '!=', 'returned')
+            ->count();
 
         if(count($book->genres) > 0) $book['genres'] = Genre::getBookGenres($book->id)[0]->genres;
         else $book['genres'] = "";
@@ -111,6 +131,10 @@ class BookController extends Controller
      */
     public function edit(Book $book)
     {
+        if(!\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+
         if(count($book->genres) > 0) $book['genres'] = Genre::getBookGenres($book->id)[0]->genres;
         else $book['genres'] = "";
 
@@ -126,6 +150,10 @@ class BookController extends Controller
      */
     public function update(Request $request, Book $book)
     {
+        if(!\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+
         $validator = Validator::make($request->all(), [
             'title' => 'required|unique:books,title,'.$book->id,
             'authors' => 'required|string|max:100',
@@ -140,6 +168,14 @@ class BookController extends Controller
 
         if ($validator->fails())
             return redirect()->back()->withErrors($validator)->withInput();
+
+        $obj = \App\Models\Transaction::bookSearchSub()
+            ->where('books.id', '=', $book->id)
+            ->first();
+
+
+        if(isset($obj->copies_used) && $request->copies_owned < $obj->copies_used)
+            return redirect()->back()->withErrors(['copies_owned' => 'Total copies less than currently used.'])->withInput();
 
         // Unlink existing image if changing or removing image.
         if((isset($request->cover_url) && $book->cover_url) || $request->cover_remove)
@@ -177,6 +213,10 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
+        if(!\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+
         if ($book->cover_url)
             unlink('media/books/'.$book->cover_url);
         return Book::deleteBook($book);
@@ -193,6 +233,10 @@ class BookController extends Controller
      */
     public function destroyWithRedirect(Book $book)
     {
+        if(!\Illuminate\Support\Facades\Auth::user()->hasRole('Librarian')){
+            return redirect()->to('/');
+        }
+        
         if ($book->cover_url)
             unlink('media/books/'.$book->cover_url);
